Moving method with sample products from servlet to ArrayListProductDao
